[FIX] html_builder, *: make the target always visible when restored
*: website

When doing undo/redo, if the operation was done in an element whose
visibility was then toggled, the option containers target may move
back inside an element that is invisible. Moreover, the changes would
also be made on an invisible element, which may confuse the user as
nothing would seem to happen when they press undo.

To fix these issues, this commit toggles the restored target visibility
when the options containers are restored on undo/redo, in order to show
it if it was invisible (or inside an invisible ancestor), thanks to the
new `on_restore_containers_handlers` resource.

Steps to reproduce:
- In website builder, drop "Text - Image" snippet.
- Click on the section and change its "Visibility" option to hide it on
  desktop
- In the "Invisible Elements" panel, toggle it to visible.
- Make a change in it.
- In the "Invisible Elements" panel, toggle it to invisible
- Undo.
=> The element where the undo took place is not visible.

task-4367641

X-original-commit: 92a766860fd616811f913ba6f276f8c02b002dbf
Co-authored by: Soukéina Bojabza <sobo@odoo.com>

Part-of: odoo/odoo#221979
Signed-off-by: Francois Georis (fge) <fge@odoo.com>
Signed-off-by: Soukéina Bojabza (sobo) <sobo@odoo.com>